feat(founders): app logic for is_founder access to Cercle Pro features
- Add centralized permissions helper (getEffectivePalier, hasCerclePro, hasPremiumFeatures, isFounder)
- Strip is_founder from public Supabase queries via toPublicPrestataire mapper
- Substitute palier with effective value before public return (no is_founder leak in SSR bundle or JSON)
- Refactor all palier gating call-sites to use helper
- Convert /annuaire from CSR to RSC to keep is_founder strictly server-side
- Cohort queries (cron stats-hebdo, benchmark stats-avancees) include founders via PostgREST .or()
- Founders bypass palier checks and get full Cercle Pro access (UI, gates, API)

Q1 (UI parity): founders displayed as Cercle Pro on badges, pastille, annuaire ranking
Q2 (out of scope): /dashboard/prestataire/abonnement intentionally untouched (deferred to PR n°3)
Q3 (cohorts): .or('palier.in.(premium,cercle_pro),is_founder.eq.true')
Q4 (privacy): is_founder never exposed in public bundle/JSON, server-trusted only
